+++ core/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./files.js";
-export * from "./gates.js";
 export * from "./routing.js";
 export * from "./serial.js";
 export * from "./steps.js";